no comment

+++ server/Fantasma.h
@@ -14,7 +14,7 @@ class Fantasma : public Personaje{
 		Enum_Personaje get_tipo();
 	
 	
-		int get_velocidad();
+		double get_velocidad();
 };
 
 #endif /*__FANTASMA_H__*/
